requirements.txt
playwright==1.49.0
pydantic==2.9.0
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
apscheduler==3.10.4
loguru==0.7.2
aiohttp==3.11.10
pandas==2.2.3
python-dotenv==1.0.1